feat: database configuration on separate library
Moved database configuration on separate library to prevent the
annoying problems when database library types are accessed on
client side.

Changes to be committed:
	modified:   src/lib/auth.ts
    - fix: changed touse dbConfig
	modified:   src/lib/db.ts
    - fix: changed touse dbConfig
    - move: getAuthDatabaseProperties => lib/dbConfig
    - move: getApiDatabaseProperties => lib/dbConfig
    - move: getTestAuthDatabaseProperties => lib/dbConfig
	new file:   src/lib/dbConfig.ts
    - feat: getAuthDatabaseProperties
    - feat: getApidatabaseProperties
    - feat: getTestAuthDatabaseProperties
	modified:   src/lib/definitions.ts
    - fix: changed touse dbConfig
	modified:   src/lib/session.ts
    - fix: changed touse dbConfig
